refactor(dotnet-broker-kafka): add createProducerServiceFile core function
