+++ plugins/dotnet-broker-kafka/src/core/index.ts
@@ -2,3 +2,4 @@ export { createStaticFileFileMap } from "./create-static-file-map";
 export { createMessageBroker } from "./create-message-broker-method";
 export { getMessageBrokerName } from "./get-message-broker-name";
 export { createConsumerServiceFile } from "./create-consumer-service-file";
+export { createProducerServiceFile } from "./create-producer-service-file";
